fix: kernel: 6.18
In the 6.18 cycle, Intel moved `struct_mutex' from the public `drm_device'
struct to `drm_i915_private'. As this was really only used by Intel to
replace the BKL (Big Kernel Lock), I think it would not be out of the
realm of reason to assume that this code was not needed at all.

Guard invocations to `struct_mutex' out on kernels >= 6.17.

Tests pending on KX-6000/6900.

Signed-off-by: Mingcong Bai <jeffbai@aosc.io>